1) Thanks to Phil @ PSU, all ports are short now. 2) idle time and grey window are cmd line options for cosignd 3) variable name changes in monster to make idle timeouts more clear 4) updated monster.8 and cosignd.8 for new changes and to document idle time out.

+++ cgi/cosigncgi.h
@@ -6,4 +6,4 @@ struct connlist {
 
 int	mkcookie( int, char * );
 void	ssl_setup( void );
-struct connlist * connlist_setup( char *, int );
+struct connlist * connlist_setup( char *, unsigned short );
